optimize logging-elasticsearch&rocketmq plugin
