+++ shenyu-plugin/shenyu-plugin-logging/shenyu-plugin-logging-elasticsearch/src/main/java/org/apache/shenyu/plugin/logging/elasticsearch/config/LogCollectConfig.java
@@ -67,6 +67,9 @@ public class LogCollectConfig {
          */
         private int maxRequestBody = 524288;
 
+        /**
+         * default 50000.
+         */
         private int bufferQueueSize = 50000;
 
         /**
