+++ shenyu-plugin/shenyu-plugin-logging/shenyu-plugin-logging-rocketmq/src/main/java/org/apache/shenyu/plugin/logging/rocketmq/LoggingRocketMQPlugin.java
@@ -24,6 +24,7 @@ import org.apache.shenyu.plugin.base.AbstractShenyuPlugin;
 import org.apache.shenyu.plugin.base.utils.HostAddressUtils;
 import org.apache.shenyu.plugin.logging.rocketmq.body.LoggingServerHttpRequest;
 import org.apache.shenyu.plugin.logging.rocketmq.body.LoggingServerHttpResponse;
+import org.apache.shenyu.plugin.logging.rocketmq.collector.DefaultLogCollector;
 import org.apache.shenyu.plugin.logging.rocketmq.entity.ShenyuRequestLog;
 import org.apache.shenyu.plugin.logging.rocketmq.utils.LogCollectConfigUtils;
 import org.apache.shenyu.plugin.logging.rocketmq.utils.LogCollectUtils;
