+++ shenyu-plugin/shenyu-plugin-logging/shenyu-plugin-logging-rocketmq/src/main/java/org/apache/shenyu/plugin/logging/rocketmq/body/LoggingServerHttpResponse.java
@@ -24,8 +24,8 @@ import org.apache.shenyu.common.utils.DateUtils;
 import org.apache.shenyu.plugin.api.context.ShenyuContext;
 import org.apache.shenyu.plugin.api.result.ShenyuResult;
 import org.apache.shenyu.plugin.api.result.ShenyuResultWrap;
-import org.apache.shenyu.plugin.logging.rocketmq.DefaultLogCollector;
-import org.apache.shenyu.plugin.logging.rocketmq.LogCollector;
+import org.apache.shenyu.plugin.logging.rocketmq.collector.DefaultLogCollector;
+import org.apache.shenyu.plugin.logging.rocketmq.collector.LogCollector;
 import org.apache.shenyu.plugin.logging.rocketmq.entity.ShenyuRequestLog;
 import org.apache.shenyu.plugin.logging.rocketmq.constant.LoggingConstant;
 import org.apache.shenyu.plugin.logging.rocketmq.utils.LogCollectConfigUtils;
