+++ shenyu-plugin/shenyu-plugin-logging/shenyu-plugin-logging-rocketmq/src/main/java/org/apache/shenyu/plugin/logging/rocketmq/config/LogCollectConfig.java
@@ -68,6 +68,9 @@ public class LogCollectConfig {
          */
         private int maxRequestBody = 524288;
 
+        /**
+         * default 50000.
+         */
         private int bufferQueueSize = 50000;
 
         /**
