+++ shenyu-plugin/shenyu-plugin-logging/shenyu-plugin-logging-rocketmq/src/test/java/org/apache/shenyu/plugin/logging/rocketmq/collector/DefaultLogCollectorTest.java
@@ -15,10 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.shenyu.plugin.logging.elasticsearch;
+package org.apache.shenyu.plugin.logging.rocketmq.collector;
 
-import org.apache.shenyu.plugin.logging.elasticsearch.elasticsearch.ElasticSearchLogCollectClient;
-import org.apache.shenyu.plugin.logging.elasticsearch.entity.ShenyuRequestLog;
+import org.apache.shenyu.plugin.logging.rocketmq.client.LogConsumeClient;
+import org.apache.shenyu.plugin.logging.rocketmq.entity.ShenyuRequestLog;
+import org.apache.shenyu.plugin.logging.rocketmq.client.RocketMQLogCollectClient;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -28,7 +29,7 @@ import java.lang.reflect.Field;
 /**
  * The Test Case For DefaultLogCollector.
  */
-public final class DefaultLogCollectorTest {
+public class DefaultLogCollectorTest {
 
     private ShenyuRequestLog shenyuRequestLog = new ShenyuRequestLog();
 
@@ -54,6 +55,6 @@ public final class DefaultLogCollectorTest {
     @Test
     public void testGetLogConsumeClient() {
         LogConsumeClient logConsumeClient = new DefaultLogCollector().getLogConsumeClient();
-        Assertions.assertEquals(ElasticSearchLogCollectClient.class, logConsumeClient.getClass());
+        Assertions.assertEquals(RocketMQLogCollectClient.class, logConsumeClient.getClass());
     }
 }
